move page folder out of the component
